+++ src/ts/BitmovinYospacePlayer.ts
@@ -50,6 +50,7 @@ import {
   YospacePlayerType,
   YospacePolicyErrorCode,
   YospaceSourceConfig,
+  AdImmunityConfig,
 } from './BitmovinYospacePlayerAPI';
 import { Logger } from './Logger';
 import { BitmovinYospaceHelper } from './BitmovinYospaceHelper';
@@ -567,6 +568,26 @@ export class BitmovinYospacePlayer implements BitmovinYospacePlayerAPI {
 
   getAspectRatio(): number {
     return this.player.getAspectRatio();
+  }
+
+  getAdImmunityConfig() {
+    return this.player.getAdImmunityConfig();
+  }
+
+  isAdImmunityActive() {
+    return this.player.isAdImmunityActive();
+  }
+
+  startAdImmunity() {
+    this.player.startAdImmunity();
+  }
+
+  endAdImmunity() {
+    this.player.endAdImmunity();
+  }
+
+  setAdImmunityConfig(options: AdImmunityConfig) {
+    this.player.setAdImmunityConfig(options);
   }
 
   readonly drm: DrmAPI;