+++ src/ts/InternalBitmovinYospacePlayer.ts
@@ -52,6 +52,8 @@ import {
 import { DefaultBitmovinYospacePlayerPolicy } from './BitmovinYospacePlayerPolicy';
 import { ArrayUtils } from 'bitmovin-player-ui/dist/js/framework/arrayutils';
 import {
+  AdImmunityConfiguredEvent,
+  AdImmunityEndedEvent,
   BitmovinYospacePlayerAPI,
   BitmovinYospacePlayerPolicy,
   CompanionAdType,
@@ -69,6 +71,8 @@ import {
   YospacePolicyErrorCode,
   YospacePolicyErrorEvent,
   YospaceSourceConfig,
+  AdImmunityStartedEvent,
+  AdImmunityConfig,
 } from './BitmovinYospacePlayerAPI';
 import { YospacePlayerError } from './YospaceError';
 import type {
@@ -147,6 +151,15 @@ export class InternalBitmovinYospacePlayer implements BitmovinYospacePlayerAPI {
   private startSent: boolean;
 
   private lastTimeChangedTime = 0;
+  private adImmunityConfig: AdImmunityConfig = {
+    duration: 0, // 0 duration = disabled
+  };
+  // Ad holiday offset for discovering upcoming ad breaks before an ad frame is shown to the user
+  private defaultAdBreakCheckOffset = 0.3;
+
+  private adImmune = false;
+  private adImmunityCountDown: number | null = null;
+  private unpauseAfterSeek = false;
 
   constructor(containerElement: HTMLElement, player: PlayerAPI, yospaceConfig: YospaceConfiguration = {}) {
     this.yospaceConfig = yospaceConfig;
@@ -251,11 +264,7 @@ export class InternalBitmovinYospacePlayer implements BitmovinYospacePlayerAPI {
               };
 
           // convert start time (relative) to an absolute time
-          if (
-            this.yospaceSourceConfig.assetType === YospaceAssetType.VOD &&
-            clonedSource.options &&
-            clonedSource.options.startOffset
-          ) {
+          if (this.yospaceSourceConfig.assetType === YospaceAssetType.VOD && clonedSource.options && clonedSource.options.startOffset) {
             clonedSource.options.startOffset = this.toAbsoluteTime(clonedSource.options.startOffset);
             Logger.log('startOffset adjusted to: ' + clonedSource.options.startOffset);
           }
@@ -393,7 +402,25 @@ export class InternalBitmovinYospacePlayer implements BitmovinYospacePlayerAPI {
       return false;
     }
 
+    // set all breaks seeked past during immunity to inactive
+    // this prevents the default player policy from redirecting
+    // the seek target
+    if (this.adImmune) {
+      const currentTime = this.player.getCurrentTime();
+
+      this.session.getAdBreaksByType(BreakType.LINEAR).forEach((adBreak) => {
+        const breakStart = this.toMagicTime(toSeconds(adBreak.getStart()));
+
+        // Check if break is being seeked past and deactivate it
+        if (breakStart > currentTime && breakStart <= time) {
+          Logger.log('[BitmovinYospacePlayer] Ad Immunity deactivated ad break during seek', adBreak);
+          adBreak.setInactive();
+        }
+      });
+    }
+
     const allowedSeekTarget = this.playerPolicy.canSeekTo(time);
+
     if (allowedSeekTarget !== time) {
       // cache original seek target
       this.cachedSeekTarget = time;
@@ -436,25 +463,21 @@ export class InternalBitmovinYospacePlayer implements BitmovinYospacePlayerAPI {
       return this.player.getDuration();
     }
 
-    return toSeconds(
-      (this.session as SessionVOD).getContentPositionForPlayhead(toMilliseconds(this.player.getDuration()))
-    );
+    return toSeconds((this.session as SessionVOD).getContentPositionForPlayhead(toMilliseconds(this.player.getDuration())));
   }
 
   /**
    * @deprecated Use {@link PlayerBufferAPI.getLevel} instead.
    */
   getVideoBufferLength(): number | null {
-    return this.buffer.getLevel(this.player.exports.BufferType.ForwardDuration, this.player.exports.MediaType.Video)
-      .level;
+    return this.buffer.getLevel(this.player.exports.BufferType.ForwardDuration, this.player.exports.MediaType.Video).level;
   }
 
   /**
    * @deprecated Use {@link PlayerBufferAPI.getLevel} instead.
    */
   getAudioBufferLength(): number | null {
-    return this.buffer.getLevel(this.player.exports.BufferType.ForwardDuration, this.player.exports.MediaType.Audio)
-      .level;
+    return this.buffer.getLevel(this.player.exports.BufferType.ForwardDuration, this.player.exports.MediaType.Audio).level;
   }
 
   getBufferedRanges(): TimeRange[] {
@@ -548,7 +571,74 @@ export class InternalBitmovinYospacePlayer implements BitmovinYospacePlayerAPI {
     this.player.setPlaybackSpeed(this.playbackSpeed);
   }
 
+  setAdImmunityConfig(config: AdImmunityConfig) {
+    this.adImmunityConfig = {
+      ...this.adImmunityConfig,
+      ...config,
+    };
+    Logger.log('[BitmovinYospacePlayer] Ad Immunity Configured:', this.adImmunityConfig);
+    this.handleYospaceEvent<AdImmunityConfiguredEvent>({
+      timestamp: Date.now(),
+      type: YospacePlayerEvent.AdImmunityConfigured,
+      config,
+    });
+  }
+
+  getAdImmunityConfig() {
+    return this.adImmunityConfig;
+  }
+
+  isAdImmunityActive() {
+    return this.adImmune;
+  }
+
+  endAdImmunity() {
+    this.endAdImmunityPeriod();
+  }
+
+  startAdImmunity() {
+    this.startAdImmunityPeriod();
+  }
+
   // Helper
+
+  private endAdImmunityPeriod() {
+    if (this.adImmunityCountDown !== null) {
+      window.clearTimeout(this.adImmunityCountDown);
+      this.adImmunityCountDown = null;
+    }
+
+    if (!this.adImmune) return;
+
+    this.adImmune = false;
+    Logger.log('[BitmovinYospacePlayer] Ad Immunity Ended');
+    this.handleYospaceEvent<AdImmunityEndedEvent>({
+      timestamp: Date.now(),
+      type: YospacePlayerEvent.AdImmunityEnded,
+    });
+  }
+
+  private startAdImmunityPeriod() {
+    // only start a timer if a duration has been configured, and ad immunity is not
+    // already active. Only enable for VOD content.
+    if (this.adImmune || this.yospaceSourceConfig.assetType !== YospaceAssetType.VOD) {
+      return;
+    } else if (this.adImmunityConfig.duration) {
+      this.adImmune = true;
+
+      this.adImmunityCountDown = window.setTimeout(() => {
+        this.endAdImmunityPeriod();
+      }, toMilliseconds(this.adImmunityConfig.duration));
+
+      Logger.log('[BitmovinYospacePlayer] Ad Immunity Started, duration', this.adImmunityConfig.duration);
+      this.handleYospaceEvent<AdImmunityStartedEvent>({
+        timestamp: Date.now(),
+        type: YospacePlayerEvent.AdImmunityStarted,
+        duration: this.adImmunityConfig.duration,
+      });
+    }
+  }
+
   private isAdActive(): boolean {
     return Boolean(this.getCurrentAd());
   }
@@ -681,10 +771,7 @@ export class InternalBitmovinYospacePlayer implements BitmovinYospacePlayerAPI {
       ...mapToYospaceCompanionAd(currentAd.getCompanionAdsByType(ResourceType.STATIC), CompanionAdType.StaticResource),
       ...mapToYospaceCompanionAd(currentAd.getCompanionAdsByType(ResourceType.HTML), CompanionAdType.HtmlResource),
       ...mapToYospaceCompanionAd(currentAd.getCompanionAdsByType(ResourceType.IFRAME), CompanionAdType.IFrameResource),
-      ...mapToYospaceCompanionAd(
-        currentAd.getCompanionAdsByType(ResourceType.UNKNOWN),
-        CompanionAdType.UnknownResource
-      ),
+      ...mapToYospaceCompanionAd(currentAd.getCompanionAdsByType(ResourceType.UNKNOWN), CompanionAdType.UnknownResource),
     ];
 
     const playerEvent = AdEventsFactory.createAdEvent(
@@ -707,11 +794,7 @@ export class InternalBitmovinYospacePlayer implements BitmovinYospacePlayerAPI {
   private onAdFinished = () => {
     const currentAd = this.getCurrentAd();
 
-    const playerEvent = AdEventsFactory.createAdEvent(
-      this.player,
-      this.player.exports.PlayerEvent.AdFinished,
-      currentAd
-    );
+    const playerEvent = AdEventsFactory.createAdEvent(this.player, this.player.exports.PlayerEvent.AdFinished, currentAd);
     this.fireEvent<AdEvent>(playerEvent);
     this.adStartedTimestamp = null;
   };
@@ -722,6 +805,8 @@ export class InternalBitmovinYospacePlayer implements BitmovinYospacePlayerAPI {
     const playerEvent = AdEventsFactory.createAdBreakEvent(this.player.exports.PlayerEvent.AdBreakFinished, adBreak);
 
     this.fireEvent<YospaceAdBreakEvent>(playerEvent);
+
+    this.startAdImmunityPeriod();
 
     if (this.cachedSeekTarget) {
       this.seek(this.cachedSeekTarget, 'yospace-ad-skipping');
@@ -756,6 +841,7 @@ export class InternalBitmovinYospacePlayer implements BitmovinYospacePlayerAPI {
       ads: ysAdBreak.getAdverts().map(AdTranslator.mapYsAdvert),
       duration: toSeconds(ysAdBreak.getDuration()),
       position: ysAdBreak.getPosition() as YospaceAdBreakPosition,
+      active: ysAdBreak.isActive(),
     };
   }
 
@@ -771,9 +857,7 @@ export class InternalBitmovinYospacePlayer implements BitmovinYospacePlayerAPI {
   }
 
   private getAdBreaksBefore(position: number): AdBreak[] {
-    return this.adParts
-      .filter((part) => part.start < position && position >= part.end)
-      .map((element) => element.adBreak);
+    return this.adParts.filter((part) => part.start < position && position >= part.end).map((element) => element.adBreak);
   }
 
   private getAdDuration(ad: Advert): number {
@@ -850,10 +934,14 @@ export class InternalBitmovinYospacePlayer implements BitmovinYospacePlayerAPI {
       this.session = null;
     }
 
+    this.endAdImmunityPeriod();
+
     if (this.dateRangeEmitter) {
       this.dateRangeEmitter.reset();
     }
 
+    // should adImmunityConfig be reset here? If yes, it
+    // would require configuration for each new video start
     this.adParts = [];
     this.adStartedTimestamp = null;
     this.cachedSeekTarget = null;
@@ -895,7 +983,7 @@ export class InternalBitmovinYospacePlayer implements BitmovinYospacePlayerAPI {
     this.fireEvent(event);
   }
 
-  private parseId3Tags(event: MetadataEvent, frames: Frame[] = []): TimedMetadata {
+  private parseId3Tags(event: MetadataEvent, frames: Frame[] = []): TimedMetadata | null {
     const charsToStr = (arr: [number]) => {
       return arr
         .filter((char) => char > 31 && char < 127)
@@ -914,8 +1002,15 @@ export class InternalBitmovinYospacePlayer implements BitmovinYospacePlayerAPI {
     }
 
     metadata.frames.forEach((frame: any) => {
-      yospaceMetadataObject[frame.key] = charsToStr(frame.data);
+      if (Array.isArray(frame.data)) {
+        yospaceMetadataObject[frame.key] = charsToStr(frame.data);
+      }
     });
+
+    if (!yospaceMetadataObject.YMID || !yospaceMetadataObject.YSEQ || !yospaceMetadataObject.YTYP || !yospaceMetadataObject.YDUR) {
+      // this was not a proper Yospace metadata but generic one and should not be passed to Yospace.
+      return null;
+    }
 
     return TimedMetadata.createFromMetadata(
       /* ymid */
@@ -931,7 +1026,7 @@ export class InternalBitmovinYospacePlayer implements BitmovinYospacePlayerAPI {
     );
   }
 
-  private mapEmsgToId3Tags(event: MetadataEvent): TimedMetadata {
+  private mapEmsgToId3Tags(event: MetadataEvent): TimedMetadata | null {
     const metadata = event.metadata as any;
     const startTime = metadata.presentationTime ? metadata.presentationTime : (this.player.getCurrentTime() as any);
 
@@ -948,18 +1043,7 @@ export class InternalBitmovinYospacePlayer implements BitmovinYospacePlayerAPI {
         return this.parseId3Tags(event, id3Frames);
       } catch (e) {
         Logger.warn(e);
-        return TimedMetadata.createFromMetadata(
-          /* ymid */
-          null,
-          /* yseq */
-          null,
-          /* ytyp */
-          null,
-          /* ydur */
-          null,
-          /* playhead */
-          toMilliseconds(startTime)
-        );
+        return null;
       }
     } else if (metadata.schemeIdUri === EmsgSchemeIdUri.V0_ID3_YOSPACE_PROPRIETARY) {
       const yospaceMetadataObject: any = {};
@@ -983,6 +1067,7 @@ export class InternalBitmovinYospacePlayer implements BitmovinYospacePlayerAPI {
       );
     } else {
       Logger.warn('Yospace integration encountered metadata that it cannot parse');
+      return null;
     }
   }
 
@@ -1022,10 +1107,13 @@ export class InternalBitmovinYospacePlayer implements BitmovinYospacePlayerAPI {
       // need to stay in the SDK as there might be AdBreak Impressions and other beacons Yospace might need to trigger.
       // These ad breaks wouldn't be visible to the user and have a duration of `0`. Yospace's recommendation for us
       // is to filter out ads with duration = 0.
-      return this.session
-        .getAdBreaksByType(BreakType.LINEAR)
-        .map((adBreak: AdBreak) => this.mapAdBreak(adBreak))
-        .filter((adBreak: YospaceAdBreak) => adBreak.duration > 0);
+      return (
+        this.session
+          .getAdBreaksByType(BreakType.LINEAR)
+          .map((adBreak: AdBreak) => this.mapAdBreak(adBreak))
+          // filter out ad breaks deactivated by ad immunity
+          .filter((adBreak: YospaceAdBreak) => adBreak.active && adBreak.duration > 0)
+      );
     },
 
     schedule: (adConfig: AdConfig) => {
@@ -1126,7 +1214,49 @@ export class InternalBitmovinYospacePlayer implements BitmovinYospacePlayerAPI {
     }
   };
 
+  private performBreakSkip(seekTarget: number) {
+    this.player.pause();
+    this.unpauseAfterSeek = true;
+
+    this.player.seek(seekTarget);
+  }
+
   private onTimeChanged = (event: TimeChangedEvent) => {
+    // the offset is an attempt to prevent the first few frames of an ad
+    // playing before the seek past the break has time to propagate
+    const adBreakCheckOffset =
+      typeof this.adImmunityConfig.adBreakCheckOffset === 'number'
+        ? this.adImmunityConfig.adBreakCheckOffset
+        : this.defaultAdBreakCheckOffset;
+    const upcomingAdBreak: AdBreak | null = this.session.getAdBreakForPlayhead(
+      toMilliseconds(event.time) + toMilliseconds(adBreakCheckOffset)
+    );
+
+    // exclude postrolls and unknown break positions from ad immunity to prevent seek loops at end of video
+    if (upcomingAdBreak?.getPosition() !== 'postroll' && upcomingAdBreak?.getPosition() !== 'unknown') {
+      // Seek past previously deactivated ad breaks
+      if (upcomingAdBreak && !upcomingAdBreak.isActive()) {
+        Logger.log('[BitmovinYospacePlayer] - Ad Immunity pausing and seeking past deactivated ad break');
+
+        this.performBreakSkip(toSeconds(upcomingAdBreak.getStart() + upcomingAdBreak.getDuration()));
+
+        // do not propagate time to the rest of the app, we want to seek past it
+        return;
+      }
+
+      // seek past and deactivate ad breaks entered during ad immunity
+      if (upcomingAdBreak && this.adImmune) {
+        upcomingAdBreak.setInactive();
+
+        Logger.log('[BitmovinYospacePlayer] - Ad Immunity pausing, seeking past and deactivating ad break');
+
+        this.performBreakSkip(toSeconds(upcomingAdBreak.getStart() + upcomingAdBreak.getDuration()));
+
+        // do not propagate time to the rest of the app, we want to seek past it
+        return;
+      }
+    }
+
     // There is an outstanding bug on Safari mobile where upon exiting an ad break,
     // our TimeChanged event "rewinds" ~12 ms. This is a temporary fix.
     // If we report this "rewind" to Yospace, it results in duplicate ad events.
@@ -1136,9 +1266,7 @@ export class InternalBitmovinYospacePlayer implements BitmovinYospacePlayerAPI {
     if (timeDifference >= 0 || timeDifference < -0.25) {
       this.session.onPlayheadUpdate(toMilliseconds(event.time));
     } else {
-      Logger.warn(
-        'Encountered a small negative TimeChanged update, not reporting to Yospace. Difference was: ' + timeDifference
-      );
+      Logger.warn('Encountered a small negative TimeChanged update, not reporting to Yospace. Difference was: ' + timeDifference);
     }
 
     // fire magic time-changed event
@@ -1170,6 +1298,12 @@ export class InternalBitmovinYospacePlayer implements BitmovinYospacePlayerAPI {
 
   private onSeeked = (event: SeekEvent) => {
     Logger.log('[BitmovinYospacePlayer] - sending YospaceAdManagement.PlayerEvent.SEEK (from Seeked player event)');
+
+    if (this.unpauseAfterSeek) {
+      this.unpauseAfterSeek = false;
+      this.play();
+    }
+
     this.session.onPlayerEvent(YsPlayerEvent.SEEK, toMilliseconds(this.player.getCurrentTime()));
 
     if (!this.suppressedEventsController.isSuppressed(this.player.exports.PlayerEvent.Seeked)) {
@@ -1210,12 +1344,15 @@ export class InternalBitmovinYospacePlayer implements BitmovinYospacePlayerAPI {
     let yospaceMetadataObject: TimedMetadata;
     if (type === 'ID3') {
       yospaceMetadataObject = this.parseId3Tags(event);
-      Logger.log('[BitmovinYospacePlayer] - sending YSPlayerEvents.METADATA ' + stringify(yospaceMetadataObject));
-      this.session.onTimedMetadata(yospaceMetadataObject);
     } else if (type === 'EMSG') {
       yospaceMetadataObject = this.mapEmsgToId3Tags(event);
+    }
+
+    if (yospaceMetadataObject) {
       Logger.log('[BitmovinYospacePlayer] - sending YSPlayerEvents.METADATA ' + stringify(yospaceMetadataObject));
       this.session.onTimedMetadata(yospaceMetadataObject);
+    } else {
+      Logger.log('[BitmovinYospacePlayer] - found metadata but does not appear to be yospace related: ' + stringify(event));
     }
   };